+++ wemake_python_styleguide/visitors/ast/modules.py
@@ -17,7 +17,13 @@ class WrongContentsVisitor(BaseNodeVisitor):
     def _is_init(self) -> bool:
         return is_stem_in_list(self.filename, [INIT])
 
-    def _is_doc_string(self, node: ast.stmt) -> bool:  # TODO: move
+    def _is_doc_string(self, node: ast.stmt) -> bool:
+        """
+        Tells whether or not the given node is a docstring.
+
+        We call docstrings any string nodes that are placed right after
+        function, class, or module definition.
+        """
         if not isinstance(node, ast.Expr):
             return False
         return isinstance(node.value, ast.Str)