+++ wemake_python_styleguide/visitors/presets/general.py
@@ -2,11 +2,15 @@
 
 from wemake_python_styleguide.visitors.ast import keywords, naming, numbers
 from wemake_python_styleguide.visitors.ast.classes import WrongClassVisitor
+from wemake_python_styleguide.visitors.ast.comparisons import (
+    ConstantComparisonVisitor,
+)
 from wemake_python_styleguide.visitors.ast.functions import (
     WrongFunctionCallVisitor,
 )
 from wemake_python_styleguide.visitors.ast.imports import WrongImportVisitor
 from wemake_python_styleguide.visitors.ast.modules import WrongContentsVisitor
+from wemake_python_styleguide.visitors.ast.order import WrongOrderVisitor
 from wemake_python_styleguide.visitors.ast.strings import WrongStringVisitor
 from wemake_python_styleguide.visitors.filenames.wrong_module_name import (
     WrongModuleNameVisitor,
@@ -24,6 +28,8 @@ GENERAL_PRESET = (
     numbers.MagicNumberVisitor,
     WrongStringVisitor,
     WrongContentsVisitor,
+    ConstantComparisonVisitor,
+    WrongOrderVisitor,
 
     # Classes:
     WrongClassVisitor,